techmarketplace/requirements.txt
Flask_SQLAlchemy==2.4.4
Flask==1.1.2
Flask_Login==0.5.0
WTForms==2.3.1
pyotp==2.3.0
pipreqs==0.4.10
Flask_WTF==0.14.3
Werkzeug==1.0.1
PyQRCode==1.2.1
itsdangerous==1.1.0
SQLAlchemy==1.3.18
Flask_Mail==0.9.1
Flask_Admin==1.5.6
requests==2.24.0
azure_storage==0.37.0
email_validator==1.1.1
flask_paranoid==0.2.0
flask_recaptcha==0.4.2
flask_session==0.3.2
mysql_connector_repackaged==0.3.1
opencensus==0.7.10
outdated==0.2.0
psutil==5.7.2
redis==3.5.3
selenium==3.141.0
vulners==1.5.5
